TestCognitoServiceFacade is moved to util subpackage
