+++ src/test/java/gov/ca/cwds/idm/GetUserTest.java
@@ -3,15 +3,15 @@ package gov.ca.cwds.idm;
 import static gov.ca.cwds.config.api.idm.Roles.CALS_ADMIN;
 import static gov.ca.cwds.config.api.idm.Roles.OFFICE_ADMIN;
 import static gov.ca.cwds.config.api.idm.Roles.STATE_ADMIN;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.ABSENT_USER_ID;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.COUNTY_ADMIN_ID;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.ERROR_USER_ID;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.STATE_ADMIN_ID;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.USER_CALS_EXTERNAL;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.USER_NO_RACFID_ID;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.USER_WITH_NO_PHONE_EXTENSION;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.USER_WITH_RACFID_AND_DB_DATA_ID;
-import static gov.ca.cwds.idm.TestCognitoServiceFacade.USER_WITH_RACFID_ID;
+import static gov.ca.cwds.idm.util.TestCognitoServiceFacade.ABSENT_USER_ID;
+import static gov.ca.cwds.idm.util.TestCognitoServiceFacade.COUNTY_ADMIN_ID;
+import static gov.ca.cwds.idm.util.TestCognitoServiceFacade.ERROR_USER_ID;
+import static gov.ca.cwds.idm.util.TestCognitoServiceFacade.STATE_ADMIN_ID;
+import static gov.ca.cwds.idm.util.TestCognitoServiceFacade.USER_CALS_EXTERNAL;
+import static gov.ca.cwds.idm.util.TestCognitoServiceFacade.USER_NO_RACFID_ID;
+import static gov.ca.cwds.idm.util.TestCognitoServiceFacade.USER_WITH_NO_PHONE_EXTENSION;
+import static gov.ca.cwds.idm.util.TestCognitoServiceFacade.USER_WITH_RACFID_AND_DB_DATA_ID;
+import static gov.ca.cwds.idm.util.TestCognitoServiceFacade.USER_WITH_RACFID_ID;
 import static gov.ca.cwds.idm.util.AssertFixtureUtils.assertNonStrict;
 
 import gov.ca.cwds.idm.util.WithMockCustomUser;
